Fix some bugs concerning simple start
When Zettelstore is started without any command line parameters, there
was an error message that it cannot connect to a certain place. This has
been fixed. Previously, the default value for the "-d" option was not
set. This is fixed a well. An previously unknown bug that prevented to
use "-d" with an absolute path is also fixed.