WebUI: simpler HTML templates
As a prerequisite for replacing the internal HTML engine, all current
templates only contain simple string variables or boolean values or list
of structs of these values.

This may break your configuration, if you edited some of those
templates.

As a side effect, an bug in v0.0.5 was fixed: setting a minimal count
value for listing tags produced a JSON list.