Add support for local references
Previously, there are references to zettel and other references. These
are interpreted as references to external material. Sometimes it is
useful to provide a reference to some internal ressources, e.g. to the
JSON encoding of a zettel or to some material that is hosted on the same
computer. These references are now interpreted as "local" references.
Similar to zettel references their URL dos not contain a scheme, user
data, a host name, nor a port. They start with a '/' character.