WebUI: change default icon for external material
This commit changes the default value of the run-time configuration key
"icon-material".

The previous SVG-based icon was difficult to handle with appropriate CSS
rules. Sometimes it was displayed at the wron size and/or at the wrong
place.

The new default value '&#8280:'. It is displayable on all tested
devices, including iPadOS. iPadOS displays some unicode character in
a "emoji-way". More important, it is much less noisy and is always
displayed at the right place and the right size.