WebUI: Allow multiple templates for new zettel
Previously, creating a new zettel was effectively cloning a special
template zettel. Now, zettel with a role "new-template" will act as
templates for creating new zettel. These zettel will be listed in the
WebUI (menu item "New" has changed from a link to a drop-down menu).

To specify the reol of the newly created zettel, a new meta data key
"new-role" was introduced.