Bug: error when no content in new zettel
To avoid creating binary content, the zettel form checked wrongly
whether the content text are is not there. This has been fixed and it is
possible to create a zettel without content, e.g. an user zettel.

This bug was introduced when it was allowed to change the meta data of
a binary zettel.