Fix some encoder when no text is given in link
Both the zmk and the native encoder generated misleading code from
a zettel when no explaining text was given in a link. In this case the
substituted text was the text of the link. This is required for the HTML
encoder, but not for the zmk encoder and for the native encoder. The
generated zmk output will not match the zmk input, for example.