Add automatic heading identifers
This applies to HTML, DJSON, and Native encoder. The text of a heading
(even if specified in Markdown) is shortened to an automatic identifier
that is emitted by above encoders. E.g. for HTML: <h2 id="abc">...</h2>.

This identifier can be used for an URI fragment to link directly to this
heading.