WebUI: initial support for list paginh
The config key 'list-page-size' controls whether all WebUI lists (list
of zettel, list of search results, possibly restricted by other
selections) are displayed in pages or not.

If set to a value greater than zero, only the specified amount of list
items is presented, together with links to the previous and to the next
page.

This will possibly be needed, if your Zettelstore contains a lot of
zettel, e.g. more that 1000.