Fix: goroutine leak in dirplace.Stop
When a dirplace is stopped, the directory service was not stopped and
removed. Instead the underlying goroutine were still running, but became
inaccessible. The leads to five new goroutine every time
a place.Reload() happened (for every running dirplace).

Now, the goroutines will finish their work successfully, when the
dirplace is stopped.