Add support for 'expert-mode'
By setting 'expert-mode' to true in runtime comfiguration, zettel with
visibility "expert" will be shown and can be updated. Most owner of
a Zettelstore will be confused by the newly introduced runtime-generated
zettel, so they are hidde nby default. This is especially useful for
Zettelstores that run in the public in 'readonly-mode'.