Make login sleep for a certain amount of time
In the current setting, login will take at least 500 msecs. If the
comparison of the given and stored credential will take sharter,
zettelstore will put itself into sleep for the difference to 500 msecs.
If the comparison takes longer, Zettelstore will slepp for approx 100
msecs.

This is the first attempt to mitigate login attacks.